requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    ./make compile-requirements
#
annotated-types==0.7.0
    # via pydantic
black==25.1.0
    # via -r requirements.in
click==8.1.8
    # via black
isort==6.0.0
    # via -r requirements.in
mypy==1.15.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.2
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pydantic==2.10.6
    # via -r requirements.in
pydantic-core==2.27.2
    # via pydantic
pyyaml==6.0.2
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   mypy
    #   pydantic
    #   pydantic-core
